*** empty log message ***
git-svn-id: http://svn.vuze.com/public/client@5534 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/core/dht/transport/udp/impl/DHTUDPPacketReply.java
@@ -39,7 +39,7 @@ DHTUDPPacketReply
 	extends PRUDPPacketReply
 {
 	private long	connection_id;
-	private int		version;
+	private byte	version;
 	private int		target_instance_id;
 	
 	public
@@ -70,7 +70,7 @@ DHTUDPPacketReply
 		
 		connection_id 	= _is.readLong();
 		
-		version			= _is.readShort();
+		version			= _is.readByte();
 			
 		DHTUDPPacket.checkVersion( version );
 	
@@ -89,7 +89,7 @@ DHTUDPPacketReply
 		return( connection_id );
 	}
 	
-	public int
+	public byte
 	getVersion()
 	{
 		return( version );
@@ -105,7 +105,7 @@ DHTUDPPacketReply
 		
 		os.writeLong( connection_id );
 		
-		os.writeShort( version );
+		os.writeByte( version );
 		
 		os.writeInt( target_instance_id );
 	}
